+++ pkg/commands/domainv1/common.go
@@ -4,14 +4,14 @@ import (
 	"fmt"
 	"io"
 
-	v1 "github.com/fastly/go-fastly/v11/fastly/domains/v1"
+	"github.com/fastly/go-fastly/v11/fastly/domainmanagement/v1/domains"
 
 	"github.com/fastly/cli/pkg/text"
 )
 
 // printSummary displays the information returned from the API in a summarised
 // format.
-func printSummary(out io.Writer, data []v1.Data) {
+func printSummary(out io.Writer, data []domains.Data) {
 	t := text.NewTable(out)
 	t.AddHeader("FQDN", "DOMAIN ID", "SERVICE ID", "CREATED AT", "UPDATED AT", "DESCRIPTION")
 	for _, d := range data {
@@ -26,7 +26,7 @@ func printSummary(out io.Writer, data []v1.Data) {
 
 // printSummary displays the information returned from the API in a verbose
 // format.
-func printVerbose(out io.Writer, data []v1.Data) {
+func printVerbose(out io.Writer, data []domains.Data) {
 	for _, d := range data {
 		fmt.Fprintf(out, "FQDN: %s\n", d.FQDN)
 		fmt.Fprintf(out, "Domain ID: %s\n", d.DomainID)