+++ pkg/commands/domainv1/delete.go
@@ -6,8 +6,7 @@ import (
 	"io"
 
 	"github.com/fastly/go-fastly/v11/fastly"
-
-	v1 "github.com/fastly/go-fastly/v11/fastly/domains/v1"
+	"github.com/fastly/go-fastly/v11/fastly/domainmanagement/v1/domains"
 
 	"github.com/fastly/cli/pkg/argparser"
 	"github.com/fastly/cli/pkg/global"
@@ -42,11 +41,11 @@ func (c *DeleteCommand) Exec(_ io.Reader, out io.Writer) error {
 		return errors.New("failed to convert interface to a fastly client")
 	}
 
-	input := &v1.DeleteInput{
+	input := &domains.DeleteInput{
 		DomainID: &c.domainID,
 	}
 
-	err := v1.Delete(context.TODO(), fc, input)
+	err := domains.Delete(context.TODO(), fc, input)
 	if err != nil {
 		c.Globals.ErrLog.AddWithContext(err, map[string]any{
 			"Domain ID": c.domainID,