+++ pkg/commands/tools/domain/status.go
@@ -6,12 +6,12 @@ import (
 	"fmt"
 	"io"
 
+	"github.com/fastly/go-fastly/v11/fastly"
+	"github.com/fastly/go-fastly/v11/fastly/domainmanagement/v1/tools/status"
+
 	"github.com/fastly/cli/pkg/argparser"
 	fsterr "github.com/fastly/cli/pkg/errors"
 	"github.com/fastly/cli/pkg/global"
-	"github.com/fastly/go-fastly/v11/fastly"
-
-	"github.com/fastly/go-fastly/v11/fastly/domains/v1/tools/status"
 )
 
 // GetDomainStatusCommand calls the Fastly API to check the availability of a domain name.